requirements.txt
# Requirements for Tomato Leaf Disease Detection System
# This file contains all Python dependencies required to run the project

# Core Deep Learning Framework
tensorflow>=2.8.0,<2.16.0    # Main deep learning library for CNN model
keras>=2.8.0                  # High-level neural network API (included with TensorFlow 2.x)

# Numerical Computing and Data Handling
numpy>=1.19.0,<1.25.0        # Fundamental package for numerical computing
pandas>=1.3.0                 # Data manipulation and analysis (optional for data processing)

# Image Processing
Pillow>=8.0.0,<10.0.0         # Python Imaging Library for image processing
opencv-python>=4.5.0          # Computer vision library (optional for advanced image processing)

# GUI Framework
# tkinter is included with Python standard library, no separate installation needed

# Data Visualization and Plotting
matplotlib>=3.3.0             # Plotting library for training curves and visualizations
seaborn>=0.11.0               # Statistical data visualization based on matplotlib

# Machine Learning Utilities
scikit-learn>=0.24.0          # Machine learning utilities for metrics and evaluation
scipy>=1.7.0                  # Scientific computing library

# Additional Utilities
tqdm>=4.62.0                  # Progress bars for loops and operations
joblib>=1.0.0                 # Lightweight pipelining for parallel computing

# Development and Testing (Optional)
pytest>=6.0.0                 # Testing framework
black>=21.0.0                 # Code formatter
flake8>=3.9.0                 # Code linting

# For saving and loading models
h5py>=3.1.0                   # HDF5 file format support for model saving

# Memory optimization (Optional)
psutil>=5.8.0                 # System and process utilities

# Note: Version ranges are specified to ensure compatibility
# If you encounter version conflicts, you may need to adjust these ranges
# For production use, consider pinning to specific versions for reproducibility

# Installation Instructions:
# 1. Create virtual environment: python -m venv tomato_env
# 2. Activate environment: 
#    - Windows: tomato_env\Scripts\activate
#    - macOS/Linux: source tomato_env/bin/activate
# 3. Install requirements: pip install -r requirements.txt

# GPU Support (Optional - for faster training):
# If you have NVIDIA GPU and want to use GPU acceleration:
# tensorflow-gpu>=2.8.0,<2.16.0  # GPU-enabled TensorFlow
# Note: Ensure you have CUDA and cuDNN installed for GPU support

# Alternative lightweight requirements (for inference only):
# If you only need to run the GUI and not train models, you can use:
# tensorflow-cpu>=2.8.0,<2.16.0  # CPU-only version (smaller download)